+++ changelog.md
@@ -1,7 +1,7 @@
 * mkchromecast (0.2.5) **unreleased**:
 
     - Volume now set to max/20 instead of max/10.
-    - This release let user linux use parec with external libraries to stream
+    - This release lets Linux users cast using  `parec` with external libraries
       instead of `ffmpeg` or `avconv`.
 
 * mkchromecast (0.2.5) **released**: 2016/05/25